HACK log: keep srd backtrace in thread local storage
Change the srd_bt_get() routine to "consume" the list of messages. Hand
ownership of previously accumulated messages to the caller, which is
supposed to release them.

Keep the 'backtrace' variable with the list of messages in thread local
storage. Since TLS variables can't have static initializers, call the
srd_bt_init() routine from the top of srd_init() for the main thread, as
well as from the top of the di_thread() routine for all child threads.

Hand decoder threads' message lists to the calling threads which join
the reaped children. This is WorkInProgress(TM). Incorporating another
thread's messages into the caller's is prepared but not implemented yet.